+++ lax/lichens/sciencerun1.py
@@ -7,9 +7,9 @@ This includes all current definitions of the cuts for the second science run
 import inspect
 import os
 import numpy as np
-from pax import units, configuration
+from pax import units
 
-from lax.lichen import Lichen, RangeLichen, ManyLichen, StringLichen
+from lax.lichen import ManyLichen, StringLichen
 from lax.lichens import sciencerun0
 from lax import __version__ as lax_version
 
@@ -27,16 +27,20 @@ class AllEnergy(ManyLichen):
 
     def __init__(self):
         self.lichen_list = [
-            FiducialCylinder1T(),
+            FiducialZOptimized(),
             InteractionExists(),
             S2Threshold(),
-            S2AreaFractionTop(),
+            InteractionPeaksBiggest(),
+            CS2AreaFractionTop(),
             S2SingleScatter(),
+            S2Width(),
             DAQVeto(),
             S1SingleScatter(),
             S2PatternLikelihood(),
-            S2Tails(),
-            InteractionPeaksBiggest()
+            KryptonMisIdS1(),
+            Flash(),
+            PosDiff(),
+            SingleElectronS2s()
         ]
 
 
@@ -51,17 +55,30 @@ class LowEnergyRn220(AllEnergy):
 
     def __init__(self):
         AllEnergy.__init__(self)
-        # Replaces Interaction exists
-        self.lichen_list[1] = S1LowEnergyRange()
 
-        # Use a simpler single scatter cut
-        self.lichen_list[4] = S2SingleScatterSimple()
+        # Customize cuts for LowE data
+        for idx, lichen in enumerate(self.lichen_list):
 
+            # Replaces InteractionExists with energy cut (tighter)
+            if lichen.name() == "CutInteractionExists":
+                self.lichen_list[idx] = S1LowEnergyRange()
+
+            # Use a simpler single scatter cut for LowE
+            if lichen.name() == "CutS2SingleScatter":
+                self.lichen_list[idx] = S2SingleScatterSimple()
+
+        # Add additional LowE cuts (that may not be tuned at HighE yet)
         self.lichen_list += [
             S1PatternLikelihood(),
-            S2Width(),
             S1MaxPMT(),
             S1AreaFractionTop(),
+            S1Width()
+        ]
+
+        # Add injection-position cuts (not for AmBe/NG)
+        self.lichen_list += [
+            S1AreaUpperInjectionFraction(),
+            S1AreaLowerInjectionFraction()
         ]
 
 
@@ -77,52 +94,115 @@ class LowEnergyBackground(LowEnergyRn220):
 
         self.lichen_list += [
             PreS2Junk(),
+            S2Tails(),  # Only for LowE background (#88)
+            MuonVeto()
         ]
+
+
+class LowEnergyAmBe(LowEnergyRn220):
+    """Select AmBe events with cs1<200 with appropriate cuts
+
+    It is the same as the LowEnergyRn220 cuts, except injection-related cuts
+    """
+
+    def __init__(self):
+        LowEnergyRn220.__init__(self)
+
+        # Remove cuts not applicable to AmBe/NG
+        self.lichen_list = [lichen for lichen in self.lichen_list
+                            if "InjectionFraction" not in lichen.name()]
+
+
+class LowEnergyNG(LowEnergyAmBe):
+    """Select NG events with cs1<200 with appropriate cuts
+
+    It is the same as the LowEnergyAmBe cuts.
+    """
+
+    def __init__(self):
+        LowEnergyAmBe.__init__(self)
 
 
 DAQVeto = sciencerun0.DAQVeto
 
+S2Tails = sciencerun0.S2Tails
 
-class S2Tails(Lichen):
-    """Check if event is in a tail of a previous S2
-    Requires S2Tail minitrees.
-    https://xecluster.lngs.infn.it/dokuwiki/doku.php?id=xenon:xenon1t:analysis:subgroup:wimphysics:s2_tails_sr0 (SR0)
-    https://xecluster.lngs.infn.it/dokuwiki/doku.php?id=xenon:xenon1t:analysis:subgroup:20170720_sr1_cut_s2_tail (SR1)
-    Contact: Daniel Coderre <daniel.coderre@physik.uni-freiburg.de>
-             Diego Ramírez <diego.ramirez@physik.uni-freiburg.de>
-    """
-    version = 1
-
-    def _process(self, df):
-        df.loc[:, self.name()] = (df['s2_over_tdiff'] < 0.025)
-        return df
+FiducialCylinder1T_TPF2dFDC = sciencerun0.FiducialCylinder1T_TPF2dFDC
 
 FiducialCylinder1T = sciencerun0.FiducialCylinder1T
 
 FiducialCylinder1p3T = sciencerun0.FiducialCylinder1p3T
 
+FiducialZOptimized = sciencerun0.FiducialZOptimized
 
-class AmBeFiducial(StringLichen):
-    """AmBe Fiducial volume cut.
-    This uses the same Z cuts as the 1T fiducial cylinder,
-    but a wider allowed range in R to maximize the number of nuclear recoils.
-    There is a third cut on the distance to the source, so that we cut away background ER.
-    Link to note:
-    https://xecluster.lngs.infn.it/dokuwiki/lib/exe/fetch.php?media=xenon:xenon1t:hogenbirk:nr_band_sr0.html
+FiducialInnerEgg = sciencerun0.FiducialInnerEgg
 
-    Contact: Erik Hogenbirk <ehogenbi@nikhef.nl>
+FV_CONFIGS = [
+    # Mass (kg), (z0, vz, p, vr2)
+    (1000, (-57.58, 31.25, 4.20, 1932.53)),
+    (1025, (-57.29, 31.65, 3.71, 1987.85)),
+    (1050, (-62.25, 33.89, 3.08, 1969.68)),
+    (1075, (-59.73, 35.58, 2.97, 1938.27)),
+    (1100, (-58.36, 36.97, 2.67, 1951.06)),
+    (1125, (-58.57, 37.36, 2.78, 1953.64)),
+    (1150, (-58.71, 37.37, 3.25, 1934.94)),
+    (1175, (-57.35, 38.17, 3.14, 1944.21)),
+    (1200, (-56.44, 39.23, 2.88, 1961.09)),
+    (1225, (-55.81, 40.30, 2.80, 1969.58)),
+    (1250, (-55.44, 40.70, 3.21, 1936.89)),
+    (1275, (-54.62, 41.19, 3.29, 1937.40)),
+    (1300, (-54.04, 42.08, 2.56, 2041.03)),
+    (1325, (-53.31, 42.52, 2.85, 2005.24)),
+    (1350, (-51.63, 43.64, 3.15, 1949.59)),
+    (1375, (-51.85, 44.07, 2.87, 2003.40)),
+    (1400, (-52.22, 43.88, 3.88, 1949.48)),
+    (1425, (-50.87, 45.22, 2.75, 2046.11)),
+    (1450, (-51.66, 43.60, 3.58, 2053.80)),
+    (1475, (-51.99, 43.92, 3.88, 2044.43)),
+    (1500, (-52.50, 43.69, 4.31, 2059.53)),
+    (1525, (-51.51, 44.67, 3.68, 2093.78)),
+    (1550, (-51.13, 45.04, 3.98, 2088.21)),
+    (1575, (-49.44, 46.67, 3.43, 2091.66)),
+    (1600, (-50.15, 45.95, 3.77, 2126.11)),
+    (1625, (-50.06, 45.99, 4.32, 2119.37)),
+    (1650, (-50.16, 45.95, 4.67, 2137.15)),
+    (1675, (-49.10, 47.05, 4.53, 2128.00)),
+    (1700, (-49.54, 46.51, 6.10, 2129.72)),
+]
 
-    Position updated to reflect correct I-Belt 1 position. Link to Note:xenon:xenon1t:analysis:dominick:sr1_ambe_check
+
+class FiducialTestEllips(StringLichen):
+    """TESTFiducial volume cut using NN 3D FDC.
+    Temporary/under development, for preliminary
+    comparisons between the different masses. For every mass
+    in the fv_config keys a FiducialTestEllips<mass> is made.
+    For more info on the construction of the FV see:
+    https://xe1t-wiki.lngs.infn.it/doku.php?id=xenon:xenon1t:analysis:sciencerun1:fiducial_volume:optimized_ellips
+    sanderb@nikhef.nl
     """
-    version = 2
-    string = "(distance_to_source < 103.5) & (-92.9 < z) & (z < -9) & (sqrt(x*x + y*y) < 42.00)"
+    version = 1
+    parameter_symbols = tuple('z0 vz p vr2'.split())
+    parameter_values = None   # Will be tuple of parameter values
+    string = "((((((z_3d_nn-@z0)**2)**0.5)/@vz)**@p)+(r_3d_nn**2/@vr2)**@p)<1"
+
+    def _process(self, df):
+        bla = dict(zip(self.parameter_symbols, self.parameter_values))
+        df.loc[:, self.name()] = df.eval(self.string,
+                                         global_dict=bla)
+        return df
 
     def pre(self, df):
-        source_position = (97, 43.5, -50)
-        df.loc[:, 'distance_to_source'] = ((source_position[0] - df['x']) ** 2 +
-                                           (source_position[1] - df['y']) ** 2 +
-                                           (source_position[2] - df['z']) ** 2) ** 0.5
+        df.loc[:, 'r_3d_nn'] = np.sqrt(df['x_3d_nn']**2 + df['y_3d_nn']**2)
         return df
+
+
+for mass, params in FV_CONFIGS:
+    name = 'FiducialTestEllips' + str(int(mass))
+    c = type(name, (FiducialTestEllips,), dict())
+    c.parameter_values = params
+    locals()[name] = c
+
+AmBeFiducial = sciencerun0.AmBeFiducial
 
 
 class NGFiducial(StringLichen):
@@ -152,9 +232,11 @@ S1MaxPMT = sciencerun0.S1MaxPMT
 
 S1PatternLikelihood = sciencerun0.S1PatternLikelihood
 
-S1SingleScatter = sciencerun0.S1SingleScatter
+S1Width = sciencerun0.S1Width
 
 S2AreaFractionTop = sciencerun0.S2AreaFractionTop
+
+CS2AreaFractionTop = sciencerun0.CS2AreaFractionTop
 
 S2SingleScatter = sciencerun0.S2SingleScatter
 
@@ -176,66 +258,39 @@ class S2PatternLikelihood(StringLichen):
 S2Threshold = sciencerun0.S2Threshold
 
 
-class S2Width(ManyLichen):
-    """S2 Width cut based on diffusion model
-    The S2 width cut compares the S2 width to what we could expect based on its depth in the detector. The inputs to
-    this are the drift velocity and the diffusion constant. The allowed variation in S2 width is greater at low
-    energy (since it is fluctuating statistically) Ref: (arXiv:1102.2865)
-    It should be applicable to data regardless of if it ER or NR;
-    above cS2 = 1e5 pe ERs the acceptance will go down due to track length effects.
-    Tune the diffusion model parameters based on pax v6.4.2 AmBe data according to note:
-    xenon:xenon1t:yuehuan:analysis:0sciencerun_s2width_update0#comparison_with_diffusion_model_cut_by_jelle_pax_v642
-    Contact: Yuehuan <weiyh@physik.uzh.ch>, Jelle <jaalbers@nikhef.nl>
+class S2Width(sciencerun0.S2Width):
+    """S2 Width cut based on diffusion model with SR1 parameters
+    See sciencerun0.py for full implementation
+    """
+    version = 6
+    diffusion_constant = 29.35 * ((units.cm)**2) / units.s
+    v_drift = 1.335 * (units.um) / units.ns
+    scg = 21.3  # s2_secondary_sc_gain in pax config
+    scw = 229.58  # s2_secondary_sc_width median
+
+
+class S1SingleScatter(sciencerun0.S1SingleScatter):
+    """S1 Single Scatter cut based on SR1 width model
+    See sciencerun0.py for full implementation
     """
     version = 3
+    s2width = S2Width
 
-    def __init__(self):
-        self.lichen_list = [self.S2WidthHigh(),
-                            self.S2WidthLow()]
 
-    @staticmethod
-    def s2_width_model(z):
-        diffusion_constant = 31.73 * ((units.cm)**2) / units.s
-        v_drift = 1.335 * (units.um) / units.ns
-        GausSigmaToR50 = 1.349
-
-        EffectivePar = 0.925
-        Sigma_0 = 229.58 * units.ns
-        return GausSigmaToR50 * np.sqrt(Sigma_0 ** 2 - EffectivePar * 2 * diffusion_constant * z / v_drift ** 3)
-
-    @staticmethod
-    def subpre(self, df):
-        df.loc[:, 'temp'] = df['s2_range_50p_area'] / S2Width.s2_width_model(df['z'])
-        return df
-
-    @staticmethod
-    def relative_s2_width_bounds(s2, kind='high'):
-        x = 0.5 * np.log10(np.clip(s2, 150, 4500 if kind == 'high' else 2500))
-        if kind == 'high':
-            return 3 - x
-        elif kind == 'low':
-            return -0.9 + x
-        raise ValueError("kind must be high or low")
-
-    class S2WidthHigh(Lichen):
-
-        def pre(self, df):
-            return S2Width.subpre(self, df)
-
-        def _process(self, df):
-            df.loc[:, self.name()] = (df.temp <= S2Width.relative_s2_width_bounds(df.s2, kind='high'))
-            return df
-
-    class S2WidthLow(RangeLichen):
-
-        def pre(self, df):
-            return S2Width.subpre(self, df)
-
-        def _process(self, df):
-            df.loc[:, self.name()] = (S2Width.relative_s2_width_bounds(df.s2, kind='low') <= df.temp)
-            return df
-
+MuonVeto = sciencerun0.MuonVeto
 
 S1AreaFractionTop = sciencerun0.S1AreaFractionTop
 
 PreS2Junk = sciencerun0.PreS2Junk
+
+KryptonMisIdS1 = sciencerun0.KryptonMisIdS1
+
+Flash = sciencerun0.Flash
+
+PosDiff = sciencerun0.PosDiff
+
+S1AreaUpperInjectionFraction = sciencerun0.S1AreaUpperInjectionFraction
+
+S1AreaLowerInjectionFraction = sciencerun0.S1AreaLowerInjectionFraction
+
+SingleElectronS2s = sciencerun0.SingleElectronS2s